support check id before update

+++ src/main/java/com/example/demo/dto/SubjectDTO.java
@@ -1,24 +1,24 @@
 package com.example.demo.dto;
 
 public class SubjectDTO {
-    int id;
+    Long id;
     String code;
     String name;
 
     SubjectDTO() {
     }
 
-    public SubjectDTO(int id, String code, String name) {
+    public SubjectDTO(Long id, String code, String name) {
         this.id = id;
         this.code = code;
         this.name = name;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
